use an iterator instead of a counter variable for info
this is slightly more efficient than the previous way of doing it because
indexing always checks for out-of-bounds errors, while iterators don't require
that. Note that the next logical step might be something like a zip:

`for (line, info) in FERRIS_ART.iter().zip(info.iter()) {`

but that would require both of them to be the same length or there would
be lines missing.